+++ src/lib/Archive/isArchiveOwner.ts
@@ -2,7 +2,7 @@ import { NoReqBody, NoReqQuery, NoResBody } from '@xylabs/sdk-api-express-ecs'
 import { XyoArchive } from '@xyo-network/sdk-xyo-client-js'
 import { Request } from 'express'
 
-import { ArchiveLocals, ArchivePathParams } from '../model'
+import { ArchiveLocals, ArchivePathParams } from '../../model'
 
 export const isRequestUserOwnerOfArchive = (req: Request, archive: XyoArchive | null | undefined): boolean => {
   // If the archive doesn't exist or have an owner, they are not the owner
