+++ src/lib/Payload/scrubBoundWitnesses.ts
