+++ src/lib/index.ts
@@ -1,16 +1,8 @@
+export * from './Archive'
 export * from './dbSdk'
-export * from './determineArchiveAccessControl'
-export * from './findByHash'
 export * from './flattenArray'
-export * from './getArchiveKeys'
 export * from './getHttpHeader'
 export * from './getQueue'
-export * from './getRequestMeta'
-export * from './isArchiveOwner'
-export * from './isPublicArchive'
-export * from './isValidArchiveName'
-export * from './prepareBoundWitnesses'
+export * from './Payload'
 export * from './Repository'
-export * from './requestCanAccessArchive'
 export * from './schema'
-export * from './scrubBoundWitnesses'
